backend/requirements.txt
anyio==3.6.1
astroid==2.11.7
backports.cached-property==1.0.2
black==22.6.0
click==8.1.3
dill==0.3.5.1
fastapi==0.79.0
flake8==4.0.1
graphql-core==3.2.1
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pydantic==1.9.1
pyflakes==2.4.0
Pygments==2.12.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
strawberry-graphql==0.121.1
tomli==2.0.1
tomlkit==0.11.1
typing_extensions==4.3.0
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.16.1
websockets==10.3
wrapt==1.14.1
